clean up serverless
