+++ bin/elasticdump
@@ -8,8 +8,7 @@ const packageData = require(path.join(__dirname, '..', 'package.json'))
 const { isUrl } = require(path.join(__dirname, '..', 'lib', 'is-url.js'))
 const ArgParser = require(path.join(__dirname, '..', 'lib', 'argv.js'))
 const versionCheck = require(path.join(__dirname, '..', 'lib', 'version-check.js'))
-require('aws-sdk/lib/maintenance_mode_message').suppress = true;
-
+require('aws-sdk/lib/maintenance_mode_message').suppress = true
 
 // For future developers.  If you add options here, be sure to add the option to test suite tests where necessary
 const defaults = {
@@ -145,7 +144,7 @@ const defaults = {
   // opensearch compatability
   'force-os-version': '7.10.2',
   // opensearch serverless
-  openSearchServerless: false,
+  openSearchServerless: false
 }
 const options = {}
