+++ lib/aws4signer.js
@@ -73,12 +73,8 @@ const aws4signer = async (esRequest, parent) => {
     esRequest.headers = Object.assign({ host: urlObj.hostname, 'Content-Type': 'application/json' }, esRequest.headers)
     esRequest.path = `${urlObj.pathname}?${urlObj.searchParams.toString()}`
 
-    const sha256 = calculateSHA256(esRequest.body)
-    esRequest.headers = {
-      'X-Amz-Content-SHA256': sha256,
-      host: urlObj.hostname,
-      'Content-Type': 'application/json',
-      ...esRequest.headers
+    if (parent.options.openSearchServerless) {
+      esRequest.headers = Object.assign({ 'X-Amz-Content-SHA256': calculateSHA256(esRequest.body) }, esRequest.headers)
     }
     return aws4.sign(esRequest, credentials)
   }
