+++ lib/transports/__es__/_base.js
@@ -35,10 +35,10 @@ class Base {
     }
   }
 
-  version (prefix, isOpenSearchServerless, callback) {
+  version (prefix, callback) {
     if (this.ESversion) { return callback() }
 
-    if (isOpenSearchServerless) {
+    if (this.parent.options.openSearchServerless) {
       this.parent.emit('debug', `opensearch serverless does not support version check, assuming major version: ${this.ESversion}`)
       this.setSearchBody()
       return callback()
