+++ lib/transports/__es__/_data.js
@@ -6,79 +6,10 @@ const _ = require('lodash')
 const { scrollResultSet, safeDecodeURIComponent } = require('./_helpers')
 
 class Data {
-  async getOpenSearchServerlessData (limit, offset, callback) {
-    let searchBodyTmp
-
-    if (this.parent.options.searchWithTemplate) {
-      searchBodyTmp = await this.renderTemplate(this.searchBody.id, this.searchBody.params)
-        .then(result => {
-          return result
-        })
-        .catch(error => {
-          throw new Error(error)
-        })
-    } else {
-      searchBodyTmp = this.searchBody
-    }
-
-    const searchBody = searchBodyTmp
-    searchBody.size = limit
-    searchBody.from = offset
-    const additionalParams = this.paramsToString(this.parent.options[`${this.options.type}-params`] || this.parent.options.params, '&')
-    // Note: OpenSearch Serverless does not support _search?scroll API. As a workaround pure _search is used
-    // List of supported endpoints: https://docs.aws.amazon.com/opensearch-service/latest/developerguide/serverless-genref.html#serverless-operations
-    const uri = `${this.base.url}/_search${additionalParams}`
-
-    const searchRequest = {
-      uri,
-      method: 'GET',
-      body: jsonParser.stringify(searchBody)
-    }
-    aws4signer(searchRequest, this.parent).then((res) => {
-      this.baseRequest(searchRequest, (err, response) => {
-        err = this.handleError(err, response)
-        if (err) {
-          return callback(err, [])
-        }
-
-        const body = jsonParser.parse(response.body, this.parent)
-
-        const hits = _.get(body, 'hits.hits', [])
-        const hitsTotal = _.get(body, 'hits.total.value', body.hits.total)
-        this.totalSearchResults = hitsTotal
-
-        if (hits.length < limit) {
-          // No more data to fetch
-          callback(null, hits)
-        } else {
-          this.getOpenSearchServerlessData(limit, offset + limit, (err, moreHits) => {
-            if (err) {
-              return callback(err, [])
-            }
-            callback(null, hits.concat(moreHits))
-          })
-        }
-      })
-    }).catch(callback)
-  }
-
   async getData (limit, offset, callback) {
     let searchRequest, uri
-    let searchBodyTmp
 
-    if (this.parent.options.searchWithTemplate) {
-      searchBodyTmp = await this.renderTemplate(this.searchBody.id, this.searchBody.params)
-        .then(result => {
-          return result
-        })
-        .catch(error => {
-          throw new Error(error)
-        })
-    } else {
-      searchBodyTmp = this.searchBody
-    }
-
-    const searchBody = searchBodyTmp
+    const searchBody = await this.searchWithTemplate(this.searchBody)
 
     if (offset >= this.totalSearchResults && this.totalSearchResults !== 0) {
       callback(null, [])
