+++ lib/transports/__es__/_template.js
@@ -65,6 +65,16 @@ class Template {
     )
   }
 
+  searchWithTemplate (searchBody) {
+    return new Promise((resolve, reject) => {
+      if (this.parent.options.searchWithTemplate) {
+        this.renderTemplate(searchBody.id, searchBody.params).then(resolve).catch(reject)
+      } else {
+        resolve(searchBody)
+      }
+    })
+  }
+
   _transformTemplates (data, type) {
     return _.reduce(data[`${type}s`], (obj, tmpl) => {
       obj[tmpl.name] = tmpl[type]
