+++ lib/transports/__es__/index.js
@@ -7,5 +7,6 @@ module.exports = {
   Setting: require('./_setting'),
   Template: require('./_template'),
   Script: require('./_script'),
-  Data: require('./_data')
+  Data: require('./_data'),
+  Serverless: require('./_serverless')
 }
