+++ lib/transports/elasticsearch.js
@@ -14,10 +14,11 @@ const {
   Policy,
   Setting,
   Template,
+  Serverless,
   Script
 } = require('./__es__')
 
-class elasticsearch extends Many(Base, Alias, Analyzer, Mapping, Policy, Setting, Template, Script, Data) {
+class elasticsearch extends Many(Base, Alias, Analyzer, Mapping, Policy, Setting, Template, Script, Data, Serverless) {
   constructor (parent, url, options) {
     super()
     this.base = parseBaseURL(url, options)
@@ -80,7 +81,7 @@ class elasticsearch extends Many(Base, Alias, Analyzer, Mapping, Policy, Setting
   // return (error, arr) where arr is an array of objects
   get (limit, offset, callback) {
     const type = this.parent.options.type
-    this.version('input', this.parent.options.openSearchServerless, err => {
+    this.version('input', err => {
       if (err) { return callback(err) }
 
       if (type === 'data') {
@@ -127,7 +128,7 @@ class elasticsearch extends Many(Base, Alias, Analyzer, Mapping, Policy, Setting
   // return (error, writes)
   set (data, limit, offset, callback) {
     const type = this.parent.options.type
-    this.version('output', this.parent.options.openSearchServerless, err => {
+    this.version('output', err => {
       if (err) { return callback(err) }
       if (type === 'data') {
         this.setData(data, limit, offset, callback)
