+++ package.json
@@ -32,7 +32,6 @@
     "aws4": "^1.12.0",
     "big.js": "^5.2.2",
     "bytes": "^3.1.2",
-    "crypto": "^1.0.1",
     "delay": "^5.0.0",
     "extends-classes": "1.0.5",
     "fast-csv": "4.3.6",
